# src/main/java/com/sumit/pack/models/Employee.java
/*
 * The Employee API
 * Creting openApi 3.0 for Spring
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.sumit.pack.models;

import java.util.Objects;
import java.util.Arrays;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;

/**
 * this is the employee model
 */
@ApiModel(description = "this is the employee model")
@JsonPropertyOrder({
  Employee.JSON_PROPERTY_ID,
  Employee.JSON_PROPERTY_NAME,
  Employee.JSON_PROPERTY_SALARY
})
@JsonTypeName("Employee")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen")
public class Employee {
  public static final String JSON_PROPERTY_ID = "id";
  private Integer id;

  public static final String JSON_PROPERTY_NAME = "name";
  private String name;

  public static final String JSON_PROPERTY_SALARY = "salary";
  private Integer salary;


  public Employee id(Integer id) {
    
    this.id = id;
    return this;
  }

   /**
   * the id of employee
   * @return id
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "the id of employee")
  @JsonProperty(JSON_PROPERTY_ID)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public Integer getId() {
    return id;
  }


  public void setId(Integer id) {
    this.id = id;
  }


  public Employee name(String name) {
    
    this.name = name;
    return this;
  }

   /**
   * name of employee
   * @return name
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "name of employee")
  @JsonProperty(JSON_PROPERTY_NAME)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getName() {
    return name;
  }


  public void setName(String name) {
    this.name = name;
  }


  public Employee salary(Integer salary) {
    
    this.salary = salary;
    return this;
  }

   /**
   * salary of employee
   * @return salary
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "salary of employee")
  @JsonProperty(JSON_PROPERTY_SALARY)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public Integer getSalary() {
    return salary;
  }


  public void setSalary(Integer salary) {
    this.salary = salary;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Employee employee = (Employee) o;
    return Objects.equals(this.id, employee.id) &&
        Objects.equals(this.name, employee.name) &&
        Objects.equals(this.salary, employee.salary);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, name, salary);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Employee {\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    salary: ").append(toIndentedString(salary)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}